requirements.txt
Flask==3.0.0
Pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
numpy==1.24.3
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0 